Update to Advanced settings
- Added the ability to change the server world in the advanced settings

- Fixed a bug with constants.safe_delete() not checking if the destination is None

+++ source/constants.py
@@ -602,6 +602,8 @@ def download_url(url: str, file_name: str, output_path: str, progress_func=None)
 
 # Will attempt to delete dir tree without error
 def safe_delete(directory: str):
+    if not directory:
+        return
     try:
         if os.path.exists(directory):
             rmtree(directory)
